@@ -1,22 +1,22 @@
 name: stream_chat_flutter_core
 homepage: https://github.com/GetStream/stream-chat-flutter
 description: Stream Chat official Flutter SDK Core. Build your own chat experience using Dart and Flutter.
-version: 3.2.0
+version: 3.3.1
 repository: https://github.com/GetStream/stream-chat-flutter
 issue_tracker: https://github.com/GetStream/stream-chat-flutter/issues
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.14.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
   collection: ^1.15.0
-  connectivity_plus: ^2.0.2
+  connectivity_plus: ^2.1.0
   flutter:
     sdk: flutter
   meta: ^1.3.0
   rxdart: ^0.27.0
-  stream_chat: ^3.2.0
+  stream_chat: ^3.3.1
 
 dev_dependencies:
   dart_code_metrics: ^4.4.0
