@@ -648,7 +648,7 @@ class Channel {
   }
 
   /// Deletes the [message] from the channel.
-  Future<EmptyResponse> deleteMessage(Message message) async {
+  Future<EmptyResponse> deleteMessage(Message message, {bool? hard}) async {
     // Directly deleting the local messages which are not yet sent to server
     if (message.status == MessageSendingStatus.sending ||
         message.status == MessageSendingStatus.failed) {
@@ -675,7 +675,7 @@ class Channel {
 
       state?.addMessage(message);
 
-      final response = await _client.deleteMessage(message.id);
+      final response = await _client.deleteMessage(message.id, hard: hard);
 
       state?.addMessage(message.copyWith(status: MessageSendingStatus.sent));
 
@@ -1466,8 +1466,6 @@ class ChannelClientState {
 
     _listenMemberRemoved();
 
-    _computeUnread();
-
     _startCleaning();
 
     _startCleaningPinnedMessages();
@@ -1489,15 +1487,6 @@ class ChannelClientState {
   }
 
   final _subscriptions = <StreamSubscription>[];
-
-  void _computeUnread() {
-    final userRead = channelState.read.firstWhereOrNull(
-      (r) => r.user.id == _channel._client.state.currentUser?.id,
-    );
-    if (userRead != null && userRead.unreadMessages > 0) {
-      unreadCount = userRead.unreadMessages;
-    }
-  }
 
   void _checkExpiredAttachmentMessages(ChannelState channelState) async {
     final expiredAttachmentMessagesId = channelState.messages
@@ -1603,7 +1592,7 @@ class ChannelClientState {
                   message.createdAt.isBefore(
                     DateTime.now().subtract(
                       const Duration(
-                        seconds: 1,
+                        seconds: 5,
                       ),
                     ),
                   ),
@@ -1663,7 +1652,11 @@ class ChannelClientState {
   void _listenMessageDeleted() {
     _subscriptions.add(_channel.on(EventType.messageDeleted).listen((event) {
       final message = event.message!;
-      addMessage(message);
+      if (event.hardDelete == true) {
+        removeMessage(message, hardDelete: true);
+      } else {
+        addMessage(message);
+      }
     }));
   }
 
@@ -1705,7 +1698,7 @@ class ChannelClientState {
       }
 
       _channelState = _channelState.copyWith(
-        messages: newMessages,
+        messages: newMessages..sort(_sortByCreatedAt),
         channel: _channelState.channel?.copyWith(
           lastMessageAt: message.createdAt,
         ),
@@ -1718,7 +1711,7 @@ class ChannelClientState {
   }
 
   /// Remove a [message] from this [channelState].
-  void removeMessage(Message message) {
+  void removeMessage(Message message, {bool hardDelete = false}) {
     final parentId = message.parentId;
     // i.e. it's a thread message
     // 1. Remove the thread message
@@ -1740,7 +1733,10 @@ class ChannelClientState {
     } else {
       // Remove regular message
       final allMessages = [...messages];
-      if (allMessages.remove(message)) {
+      if (hardDelete) {
+        allMessages.removeWhere((e) => e.id == message.id);
+        _channelState = _channelState.copyWith(messages: allMessages);
+      } else if (allMessages.remove(message)) {
         _channelState = _channelState.copyWith(messages: allMessages);
       }
     }
@@ -1843,15 +1839,34 @@ class ChannelClientState {
   /// Channel read list as a stream.
   Stream<List<Read>> get readStream => channelStateStream.map((cs) => cs.read);
 
-  final BehaviorSubject<int> _unreadCountController = BehaviorSubject.seeded(0);
+  bool _isCurrentUserRead(Read read) =>
+      read.user.id == _channel._client.state.currentUser!.id;
 
-  set unreadCount(int value) => _unreadCountController.add(value);
+  /// Channel read for the logged in user.
+  Read? get currentUserRead => read.firstWhereOrNull(_isCurrentUserRead);
+
+  /// Channel read for the logged in user as a stream.
+  Stream<Read?> get currentUserReadStream =>
+      readStream.map((read) => read.firstWhereOrNull(_isCurrentUserRead));
 
   /// Unread count getter as a stream.
-  Stream<int> get unreadCountStream => _unreadCountController.stream.distinct();
+  Stream<int> get unreadCountStream =>
+      currentUserReadStream.map((read) => read?.unreadMessages ?? 0);
 
   /// Unread count getter.
-  int get unreadCount => _unreadCountController.value;
+  int get unreadCount => currentUserRead?.unreadMessages ?? 0;
+
+  /// Setter for unread count.
+  set unreadCount(int count) {
+    final reads = [..._channelState.read];
+    final currentUserReadIndex = reads.indexWhere(_isCurrentUserRead);
+
+    if (currentUserReadIndex < 0) return;
+
+    reads[currentUserReadIndex] =
+        reads[currentUserReadIndex].copyWith(unreadMessages: count);
+    _channelState = _channelState.copyWith(read: reads);
+  }
 
   bool _countMessageAsUnread(Message message) {
     final userId = _channel.client.state.currentUser?.id;
@@ -1860,10 +1875,13 @@ class ChannelClientState {
               (m) => m.user.id == message.user?.id,
             ) !=
             null;
+    final isThreadMessage = message.parentId != null;
+
     return !message.silent &&
         !message.shadowed &&
         message.user?.id != userId &&
-        !userIsMuted;
+        !userIsMuted &&
+        !isThreadMessage;
   }
 
   /// Update threads with updated information about messages.
@@ -2108,7 +2126,6 @@ class ChannelClientState {
   /// Call this method to dispose this object.
   void dispose() {
     _debouncedUpdatePersistenceChannelState.cancel();
-    _unreadCountController.close();
     _retryQueue.dispose();
     _subscriptions.forEach((s) => s.cancel());
     _channelStateController.close();
