@@ -44,10 +44,6 @@ final _levelEmojiMapper = {
   Level.SEVERE: '🚨',
 };
 
-final _userAgent = 'stream-chat-dart-client-'
-    '${CurrentPlatform.name}-'
-    '${PACKAGE_VERSION.split('+')[0]}';
-
 /// The official Dart client for Stream Chat,
 /// a service for building chat applications.
 /// This library can be used on any Dart project and on both mobile and web apps
@@ -86,7 +82,7 @@ class StreamChatClient {
       location: location,
       connectTimeout: connectTimeout,
       receiveTimeout: receiveTimeout,
-      headers: {'X-Stream-Client': _userAgent},
+      headers: {'X-Stream-Client': defaultUserAgent},
     );
 
     _chatApi = chatApi ??
@@ -106,7 +102,7 @@ class StreamChatClient {
           tokenManager: _tokenManager,
           handler: handleEvent,
           logger: detachedLogger('🔌'),
-          queryParameters: {'X-Stream-Client': _userAgent},
+          queryParameters: {'X-Stream-Client': defaultUserAgent},
         );
 
     _retryPolicy = retryPolicy ??
@@ -130,6 +126,14 @@ class StreamChatClient {
   set chatPersistenceClient(ChatPersistenceClient? value) {
     _originalChatPersistenceClient = value;
   }
+
+  /// Default user agent for all requests
+  static String defaultUserAgent = 'stream-chat-dart-client-'
+      '${CurrentPlatform.name}-'
+      '${PACKAGE_VERSION.split('+')[0]}';
+
+  /// Additionals headers for all requests
+  static Map<String, Object?> additionalHeaders = {};
 
   ChatPersistenceClient? _originalChatPersistenceClient;
 
@@ -391,6 +395,9 @@ class StreamChatClient {
   }
 
   void _handleHealthCheckEvent(Event event) {
+    final user = event.me;
+    if (user != null) state.currentUser = user;
+
     final connectionId = event.connectionId;
     if (connectionId != null) {
       _connectionIdManager.setConnectionId(connectionId);
@@ -1209,8 +1216,14 @@ class StreamChatClient {
       );
 
   /// Deletes the given message
-  Future<EmptyResponse> deleteMessage(String messageId) =>
-      _chatApi.message.deleteMessage(messageId);
+  Future<EmptyResponse> deleteMessage(String messageId, {bool? hard}) async {
+    final response =
+        await _chatApi.message.deleteMessage(messageId, hard: hard);
+    if (hard == true) {
+      await _chatPersistenceClient?.deleteMessageById(messageId);
+    }
+    return response;
+  }
 
   /// Get a message by [messageId]
   Future<GetMessageResponse> getMessage(String messageId) =>
@@ -1302,6 +1315,10 @@ class StreamChatClient {
           'pinned': false,
         },
       );
+
+  /// Get OpenGraph data of the given [url].
+  Future<OGAttachmentResponse> enrichUrl(String url) =>
+      _chatApi.general.enrichUrl(url);
 
   /// Closes the [_ws] connection and resets the [state]
   /// If [flushChatPersistence] is true the client deletes all offline
@@ -1428,6 +1445,7 @@ class ClientState {
         .listen((Event event) async {
       final eventChannel = event.channel!;
       await _client.chatPersistenceClient?.deleteChannels([eventChannel.cid]);
+      channels[eventChannel.cid]?.dispose();
       channels = channels..remove(eventChannel.cid);
     }));
   }
