@@ -1,18 +1,13 @@
 import 'dart:async';
 
 import 'package:collection/collection.dart';
-import 'package:logging/logging.dart';
 import 'package:rxdart/rxdart.dart';
-import 'package:stream_chat/src/client/channel.dart';
 import 'package:stream_chat/src/client/retry_policy.dart';
-import 'package:stream_chat/src/core/error/error.dart';
-import 'package:stream_chat/src/core/models/message.dart';
-import 'package:stream_chat/src/event_type.dart';
 import 'package:stream_chat/stream_chat.dart';
 
-/// The retry queue associated to a channel
+/// The retry queue associated to a channel.
 class RetryQueue {
-  /// Instantiate a new RetryQueue object
+  /// Instantiate a new RetryQueue object.
   RetryQueue({
     required this.channel,
     this.logger,
@@ -22,13 +17,13 @@ class RetryQueue {
     _listenFailedEvents();
   }
 
-  /// The channel of this queue
+  /// The channel of this queue.
   final Channel channel;
 
-  /// The client associated with this [channel]
+  /// The client associated with this [channel].
   final StreamChatClient client;
 
-  /// The logger associated to this queue
+  /// The logger associated to this queue.
   final Logger? logger;
 
   late final RetryPolicy _retryPolicy;
@@ -68,17 +63,18 @@ class RetryQueue {
     }).addTo(_compositeSubscription);
   }
 
-  /// Add a list of messages
+  /// Add a list of messages.
   void add(List<Message> messages) {
     if (messages.isEmpty) return;
-    if (_messageQueue.containsAllMessage(messages)) return;
+    if (!_messageQueue.containsAllMessage(messages)) {
+      logger?.info('Adding ${messages.length} messages');
+      final messageList = _messageQueue.toList();
+      // we should not add message if already available in the queue
+      _messageQueue.addAll(messages.where(
+        (it) => !messageList.any((m) => m.id == it.id),
+      ));
+    }
 
-    logger?.info('Adding ${messages.length} messages');
-    final messageList = _messageQueue.toList();
-    // we should not add message if already available in the queue
-    _messageQueue.addAll(messages.where(
-      (it) => !messageList.any((m) => m.id == it.id),
-    ));
     _startRetrying();
   }
 
@@ -90,17 +86,21 @@ class RetryQueue {
     while (_messageQueue.isNotEmpty) {
       logger?.info('${_messageQueue.length} messages remaining in the queue');
       final message = _messageQueue.first;
-      await _runAndRetry(message);
+      final succeeded = await _runAndRetry(message);
+      if (!succeeded) {
+        _messageQueue.toList().forEach(_sendFailedEvent);
+        break;
+      }
     }
     _isRetrying = false;
   }
 
-  Future<void> _runAndRetry(Message message) async {
+  Future<bool> _runAndRetry(Message message) async {
     var attempt = 1;
 
     final maxAttempt = _retryPolicy.maxRetryAttempts;
     // early return in case maxAttempt is less than 0
-    if (attempt > maxAttempt) return;
+    if (attempt > maxAttempt) return false;
 
     // ignore: literal_only_boolean_expressions
     while (true) {
@@ -109,8 +109,13 @@ class RetryQueue {
         await _retryMessage(message);
         logger?.info('Message (${message.id}) sent successfully');
         _messageQueue.removeMessage(message);
-        break;
-      } on StreamChatError catch (e) {
+        return true;
+      } catch (e) {
+        if (e is! StreamChatNetworkError || !e.isRetriable) {
+          _messageQueue.removeMessage(message);
+          _sendFailedEvent(message);
+          return true;
+        }
         // retry logic
         final maxAttempt = _retryPolicy.maxRetryAttempts;
         if (attempt < maxAttempt) {
@@ -143,16 +148,9 @@ class RetryQueue {
           _sendFailedEvent(message);
           break;
         }
-      } catch (e) {
-        logger?.info(
-          'API call failed due to unknown error (attempt $attempt). '
-          'Giving up for now, will retry when connection recovers. '
-          'Error was $e',
-        );
-        _sendFailedEvent(message);
-        break;
       }
     }
+    return false;
   }
 
   void _sendFailedEvent(Message message) {
@@ -177,10 +175,10 @@ class RetryQueue {
     }
   }
 
-  /// Whether our [_messageQueue] has messages or not
+  /// Whether our [_messageQueue] has messages or not.
   bool get hasMessages => _messageQueue.isNotEmpty;
 
-  /// Call this method to dispose this object
+  /// Call this method to dispose this object.
   void dispose() {
     _messageQueue.clear();
     _compositeSubscription.dispose();
