@@ -4,6 +4,7 @@ import 'package:mocktail/mocktail.dart';
 import 'package:stream_chat/src/core/api/responses.dart';
 import 'package:stream_chat/src/core/error/error.dart';
 import 'package:stream_chat/src/core/http/connection_id_manager.dart';
+import 'package:stream_chat/src/core/http/interceptor/additional_headers_interceptor.dart';
 import 'package:stream_chat/src/core/http/interceptor/auth_interceptor.dart';
 import 'package:stream_chat/src/core/http/interceptor/connection_id_interceptor.dart';
 import 'package:stream_chat/src/core/http/interceptor/logging_interceptor.dart';
@@ -48,12 +49,23 @@ void main() {
     return dioError;
   }
 
+  test('UserAgentInterceptor should be added', () {
+    const apiKey = 'api-key';
+    final client = StreamHttpClient(apiKey);
+
+    expect(
+        client.httpClient.interceptors
+            .whereType<AdditionalHeadersInterceptor>()
+            .length,
+        1);
+  });
+
   test('AuthInterceptor should be added if tokenManager is provided', () {
     const apiKey = 'api-key';
     final client = StreamHttpClient(apiKey, tokenManager: TokenManager());
 
-    expect(client.httpClient.interceptors.length, 1);
-    expect(client.httpClient.interceptors.first, isA<AuthInterceptor>());
+    expect(
+        client.httpClient.interceptors.whereType<AuthInterceptor>().length, 1);
   });
 
   test(
@@ -65,10 +77,11 @@ void main() {
         connectionIdManager: ConnectionIdManager(),
       );
 
-      expect(client.httpClient.interceptors.length, 1);
       expect(
-        client.httpClient.interceptors.first,
-        isA<ConnectionIdInterceptor>(),
+        client.httpClient.interceptors
+            .whereType<ConnectionIdInterceptor>()
+            .length,
+        1,
       );
     },
   );
@@ -80,10 +93,9 @@ void main() {
       logger: Logger('test-logger'),
     );
 
-    expect(client.httpClient.interceptors.length, 1);
     expect(
-      client.httpClient.interceptors.first,
-      isA<LoggingInterceptor>(),
+      client.httpClient.interceptors.whereType<LoggingInterceptor>().length,
+      1,
     );
   });
 
@@ -109,27 +121,6 @@ void main() {
     } catch (_) {}
 
     verify(() => logger.severe(any())).called(greaterThan(0));
-  });
-
-  test('`.lock` should lock the dio client', () async {
-    final client = StreamHttpClient('api-key');
-    expect(client.httpClient.interceptors.requestLock.locked, isFalse);
-    client.lock();
-    expect(client.httpClient.interceptors.requestLock.locked, isTrue);
-  });
-
-  test('`.unlock` should unlock the dio client', () async {
-    final client = StreamHttpClient('api-key');
-    expect(client.httpClient.interceptors.requestLock.locked, isFalse);
-    client.lock();
-    expect(client.httpClient.interceptors.requestLock.locked, isTrue);
-    client.unlock();
-    expect(client.httpClient.interceptors.requestLock.locked, isFalse);
-  });
-
-  test('`.clear` should clear and unlock the dio client', () async {
-    final client = StreamHttpClient('api-key')..clear();
-    expect(client.httpClient.interceptors.requestLock.locked, isFalse);
   });
 
   test('`.close` should close the dio client', () async {
