@@ -6,6 +6,15 @@ import 'package:stream_chat_flutter/src/stream_svg_icon.dart';
 import 'package:stream_chat_flutter/src/theme/themes.dart';
 import 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
 
+/// Widget builder for attachment actions modal
+/// [defaultActionsModal] is the default [AttachmentActionsModal] config
+/// Use [defaultActionsModal.copyWith] to easily customize it
+typedef AttachmentActionsBuilder = Widget Function(
+  BuildContext context,
+  Attachment attachment,
+  AttachmentActionsModal defaultActionsModal,
+);
+
 /// Header/AppBar widget for media display screen
 class GalleryHeader extends StatelessWidget implements PreferredSizeWidget {
   /// Creates a channel header
@@ -21,6 +30,7 @@ class GalleryHeader extends StatelessWidget implements PreferredSizeWidget {
     this.userName = '',
     this.sentAt = '',
     this.backgroundColor,
+    this.attachmentActionsModalBuilder,
   })  : preferredSize = const Size.fromHeight(kToolbarHeight),
         super(key: key);
 
@@ -54,6 +64,11 @@ class GalleryHeader extends StatelessWidget implements PreferredSizeWidget {
 
   /// The background color of this [GalleryHeader].
   final Color? backgroundColor;
+
+  /// Widget builder for attachment actions modal
+  /// [defaultActionsModal] is the default [AttachmentActionsModal] config
+  /// Use [defaultActionsModal.copyWith] to easily customize it
+  final AttachmentActionsBuilder? attachmentActionsModalBuilder;
 
   @override
   Widget build(BuildContext context) {
@@ -123,17 +138,26 @@ class GalleryHeader extends StatelessWidget implements PreferredSizeWidget {
     final galleryHeaderThemeData =
         StreamChatTheme.of(context).galleryHeaderTheme;
 
+    final defaultModal = AttachmentActionsModal(
+      message: message,
+      currentIndex: currentIndex,
+      onShowMessage: onShowMessage,
+    );
+
+    final effectiveModal = attachmentActionsModalBuilder?.call(
+          context,
+          message.attachments[currentIndex],
+          defaultModal,
+        ) ??
+        defaultModal;
+
     final result = await showDialog(
       useRootNavigator: false,
       context: context,
       barrierColor: galleryHeaderThemeData.bottomSheetBarrierColor,
       builder: (context) => StreamChannel(
         channel: channel,
-        child: AttachmentActionsModal(
-          message: message,
-          currentIndex: currentIndex,
-          onShowMessage: onShowMessage,
-        ),
+        child: effectiveModal,
       ),
     );
 
