@@ -4,7 +4,6 @@ import 'dart:math';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:collection/collection.dart';
 import 'package:file_picker/file_picker.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
@@ -14,19 +13,18 @@ import 'package:stream_chat_flutter/src/commands_overlay.dart';
 import 'package:stream_chat_flutter/src/emoji/emoji.dart';
 import 'package:stream_chat_flutter/src/emoji_overlay.dart';
 import 'package:stream_chat_flutter/src/extension.dart';
-import 'package:stream_chat_flutter/src/message_list_view.dart';
 import 'package:stream_chat_flutter/src/multi_overlay.dart';
 import 'package:stream_chat_flutter/src/quoted_message_widget.dart';
-import 'package:stream_chat_flutter/src/stream_chat_theme.dart';
-import 'package:stream_chat_flutter/src/stream_svg_icon.dart';
 import 'package:stream_chat_flutter/src/user_mentions_overlay.dart';
 import 'package:stream_chat_flutter/src/video_service.dart';
 import 'package:stream_chat_flutter/src/video_thumbnail_image.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
-import 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
 import 'package:video_compress/video_compress.dart';
 
 export 'package:video_compress/video_compress.dart' show VideoQuality;
+
+/// A function that returns true if the message is valid and can be sent.
+typedef MessageValidator = bool Function(Message message);
 
 /// A callback that can be passed to [MessageInput.onError].
 ///
@@ -42,13 +40,14 @@ typedef ErrorListener = void Function(
 ///
 /// This callback should not throw.
 ///
-/// It exists merely for showing custom error, and should not be used otherwise.
+/// It exists merely for showing a custom error, and should not be used
+/// otherwise.
 typedef AttachmentLimitExceedListener = void Function(
   int limit,
   String error,
 );
 
-/// Builder for attachment thumbnails
+/// Builder for attachment thumbnails.
 typedef AttachmentThumbnailBuilder = Widget Function(
   BuildContext,
   Attachment,
@@ -77,8 +76,8 @@ typedef ActionButtonBuilder = Widget Function(
   IconButton defaultActionButton,
 );
 
-/// Widget builder for widgets that require may required data from the
-/// [MessageInputController]
+/// Widget builder for widgets that may require data from the
+/// [MessageInputController].
 typedef MessageRelatedBuilder = Widget Function(
   BuildContext context,
   MessageInputController messageInputController,
@@ -91,7 +90,7 @@ typedef AttachmentsPickerBuilder = Widget Function(
   StreamAttachmentPicker defaultPicker,
 );
 
-/// Location for actions on the [MessageInput]
+/// Location for actions on the [MessageInput].
 enum ActionsLocation {
   /// Align to left
   left,
@@ -106,7 +105,7 @@ enum ActionsLocation {
   rightInside,
 }
 
-/// Default attachments for widget
+/// Default attachments for widget.
 enum DefaultAttachmentTypes {
   /// Image Attachment
   image,
@@ -118,7 +117,7 @@ enum DefaultAttachmentTypes {
   file,
 }
 
-/// Available locations for the sendMessage button relative to the textField
+/// Available locations for the `sendMessage` button relative to the textField.
 enum SendButtonLocation {
   /// inside the textField
   inside,
@@ -131,17 +130,17 @@ const _kMinMediaPickerSize = 360.0;
 
 const _kDefaultMaxAttachmentSize = 20971520; // 20MB in Bytes
 
-/// Inactive state
+/// Inactive state:
 ///
 /// ![screenshot](https://raw.githubusercontent.com/GetStream/stream-chat-flutter/master/packages/stream_chat_flutter/screenshots/message_input.png)
 /// ![screenshot](https://raw.githubusercontent.com/GetStream/stream-chat-flutter/master/packages/stream_chat_flutter/screenshots/message_input_paint.png)
 ///
-/// Focused state
+/// Focused state:
 ///
 /// ![screenshot](https://raw.githubusercontent.com/GetStream/stream-chat-flutter/master/packages/stream_chat_flutter/screenshots/message_input2.png)
 /// ![screenshot](https://raw.githubusercontent.com/GetStream/stream-chat-flutter/master/packages/stream_chat_flutter/screenshots/message_input2_paint.png)
 ///
-/// Widget used to enter the message and add attachments
+/// Widget used to enter a message and add attachments:
 ///
 /// ```dart
 /// class ChannelPage extends StatelessWidget {
@@ -176,20 +175,16 @@ const _kDefaultMaxAttachmentSize = 20971520; // 20MB in Bytes
 /// as the bottom widget.
 ///
 /// The widget renders the ui based on the first ancestor of
-/// type [StreamChatTheme].
-/// Modify it to change the widget appearance.
+/// type [StreamChatTheme]. Modify it to change the widget appearance.
 class MessageInput extends StatefulWidget {
   /// Instantiate a new MessageInput
   const MessageInput({
     Key? key,
     this.onMessageSent,
     this.preMessageSending,
-    this.parentMessage,
-    this.editMessage,
     this.maxHeight = 150,
     this.keyboardType = TextInputType.multiline,
     this.disableAttachments = false,
-    this.initialMessage,
     this.messageInputController,
     this.actions = const [],
     this.actionsLocation = ActionsLocation.left,
@@ -218,76 +213,67 @@ class MessageInput extends StatefulWidget {
     this.mentionAllAppUsers = false,
     this.attachmentsPickerBuilder,
     this.sendButtonBuilder,
-  })  : assert(
-          initialMessage == null || editMessage == null,
-          "Can't provide both `initialMessage` and `editMessage`",
-        ),
-        super(key: key);
+    this.shouldKeepFocusAfterMessage,
+    this.validator = _defaultValidator,
+    this.restorationId,
+  }) : super(key: key);
 
-  /// List of options for showing overlays
+  /// List of options for showing overlays.
   final List<OverlayOptions> customOverlays;
 
-  /// Message to edit
-  final Message? editMessage;
-
-  /// Video quality to use when compressing the videos
+  /// Video quality to use when compressing the videos.
   final VideoQuality compressedVideoQuality;
 
-  /// Frame rate to use when compressing the videos
+  /// Frame rate to use when compressing the videos.
   final int compressedVideoFrameRate;
 
-  /// Max attachment size in bytes
-  /// Defaults to 20 MB
-  /// do not set it if you're using our default CDN
+  /// Max attachment size in bytes:
+  /// - Defaults to 20 MB
+  /// - Do not set it if you're using our default CDN
   final int maxAttachmentSize;
 
-  /// Message to start with
-  final Message? initialMessage;
-
-  /// Function called after sending the message
+  /// Function called after sending the message.
   final void Function(Message)? onMessageSent;
 
-  /// Function called right before sending the message
-  /// Use this to transform the message
+  /// Function called right before sending the message.
+  ///
+  /// Use this to transform the message.
   final FutureOr<Message> Function(Message)? preMessageSending;
 
-  /// Parent message in case of a thread
-  final Message? parentMessage;
-
-  /// Maximum Height for the TextField to grow before it starts scrolling
+  /// Maximum Height for the TextField to grow before it starts scrolling.
   final double maxHeight;
 
-  /// The keyboard type assigned to the TextField
+  /// The keyboard type assigned to the TextField.
   final TextInputType keyboardType;
 
-  /// If true the attachments button will not be displayed
+  /// If true the attachments button will not be displayed.
   final bool disableAttachments;
 
-  /// Use this property to hide/show the commands button
+  /// Use this property to hide/show the commands button.
   final bool showCommandsButton;
 
-  /// Hide send as dm checkbox
+  /// Hide send as dm checkbox.
   final bool hideSendAsDm;
 
-  /// The text controller of the TextField
+  /// The text controller of the TextField.
   final MessageInputController? messageInputController;
 
-  /// List of action widgets
+  /// List of action widgets.
   final List<Widget> actions;
 
-  /// The location of the custom actions
+  /// The location of the custom actions.
   final ActionsLocation actionsLocation;
 
-  /// Map that defines a thumbnail builder for an attachment type
+  /// Map that defines a thumbnail builder for an attachment type.
   final Map<String, AttachmentThumbnailBuilder>? attachmentThumbnailBuilders;
 
-  /// The focus node associated to the TextField
+  /// The focus node associated to the TextField.
   final FocusNode? focusNode;
 
-  ///
+  /// The message that is being quoted.
   final Message? quotedMessage;
 
-  ///
+  /// Callback invoked when the quoted message is cleared.
   final VoidCallback? onQuotedMessageCleared;
 
   /// The location of the send button
@@ -342,6 +328,19 @@ class MessageInput extends StatefulWidget {
   /// Builder for creating send button
   final MessageRelatedBuilder? sendButtonBuilder;
 
+  /// Defines if the [MessageInput] loses focuses after a message is sent.
+  /// The default behaviour keeps focus until a command is enabled.
+  final bool? shouldKeepFocusAfterMessage;
+
+  /// A callback function that validates the message.
+  final MessageValidator validator;
+
+  /// Restoration ID to save and restore the state of the MessageInput.
+  final String? restorationId;
+
+  static bool _defaultValidator(Message message) =>
+      message.text?.isNotEmpty == true || message.attachments.isNotEmpty;
+
   @override
   MessageInputState createState() => MessageInputState();
 
@@ -358,39 +357,76 @@ class MessageInput extends StatefulWidget {
 }
 
 /// State of [MessageInput]
-class MessageInputState extends State<MessageInput> {
+class MessageInputState extends State<MessageInput>
+    with RestorationMixin<MessageInput> {
   final _imagePicker = ImagePicker();
   late final _focusNode = widget.focusNode ?? FocusNode();
   bool _inputEnabled = true;
-  bool _commandEnabled = false;
+  bool get _commandEnabled => _effectiveController.value.command != null;
   bool _showCommandsOverlay = false;
   bool _showMentionsOverlay = false;
 
-  Command? _chosenCommand;
   bool _actionsShrunk = false;
   bool _openFilePickerSection = false;
-
-  /// The editing controller passed to the input TextField
-  late final MessageInputController messageInputController =
-      widget.messageInputController ?? MessageInputController();
 
   late StreamChatThemeData _streamChatTheme;
   late MessageInputThemeData _messageInputTheme;
 
-  bool get _hasQuotedMessage => widget.quotedMessage != null;
+  bool get _hasQuotedMessage =>
+      _effectiveController.value.quotedMessage != null;
 
-  bool get _messageIsPresent => messageInputController.text.trim().isNotEmpty;
+  bool get _isEditing =>
+      _effectiveController.value.status != MessageSendingStatus.sending;
+
+  RestorableMessageInputController? _controller;
+  MessageInputController get _effectiveController =>
+      widget.messageInputController ?? _controller!.value;
+
+  void _createLocalController([Message? message]) {
+    assert(_controller == null, '');
+    _controller = RestorableMessageInputController(message: message);
+    print('_controller?.value: ${_controller?.value}');
+  }
+
+  void _registerController() {
+    assert(_controller != null, '');
+    registerForRestoration(_controller!, 'messageInputController');
+  }
 
   @override
   void initState() {
     super.initState();
-    if (widget.editMessage != null || widget.initialMessage != null) {
-      _parseExistingMessage(widget.editMessage ?? widget.initialMessage!);
+    if (widget.messageInputController == null) {
+      _createLocalController();
+      print('_controller?.value: ${_controller?.value}');
     }
-    messageInputController.textEditingController
-        .addListener(_onChangedDebounced);
+    _effectiveController.textEditingController.addListener(_onChangedDebounced);
     _focusNode.addListener(_focusNodeListener);
   }
+
+  @override
+  void didUpdateWidget(covariant MessageInput oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (widget.messageInputController == null &&
+        oldWidget.messageInputController != null) {
+      _createLocalController(oldWidget.messageInputController!.value);
+    } else if (widget.messageInputController != null &&
+        oldWidget.messageInputController == null) {
+      unregisterFromRestoration(_controller!);
+      _controller!.dispose();
+      _controller = null;
+    }
+  }
+
+  @override
+  void restoreState(RestorationBucket? oldBucket, bool initialRestore) {
+    if (_controller != null) {
+      _registerController();
+    }
+  }
+
+  @override
+  String? get restorationId => widget.restorationId;
 
   void _focusNodeListener() {
     if (_focusNode.hasFocus) {
@@ -430,9 +466,9 @@ class MessageInputState extends State<MessageInput> {
 
   @override
   Widget build(BuildContext context) {
-    Widget child = ValueListenableBuilder<Message>(
-      valueListenable: messageInputController,
-      builder: (context, value, wid) => DecoratedBox(
+    Widget child = MessageValueListenableBuilder(
+      valueListenable: _effectiveController,
+      builder: (context, value, _) => DecoratedBox(
         decoration: BoxDecoration(
           color: _messageInputTheme.inputBackgroundColor,
         ),
@@ -479,7 +515,8 @@ class MessageInputState extends State<MessageInput> {
                   padding: const EdgeInsets.symmetric(vertical: 8),
                   child: _buildTextField(context),
                 ),
-                if (widget.parentMessage != null && !widget.hideSendAsDm)
+                if (_effectiveController.value.parentId != null &&
+                    !widget.hideSendAsDm)
                   Padding(
                     padding: const EdgeInsets.only(
                       right: 12,
@@ -495,7 +532,7 @@ class MessageInputState extends State<MessageInput> {
         ),
       ),
     );
-    if (widget.editMessage == null) {
+    if (_isEditing) {
       child = Material(
         elevation: 8,
         child: child,
@@ -512,12 +549,12 @@ class MessageInputState extends State<MessageInput> {
         ),
         OverlayOptions(
           visible: _focusNode.hasFocus &&
-              messageInputController.text.isNotEmpty &&
-              messageInputController.baseOffset > 0 &&
-              messageInputController.text
+              _effectiveController.text.isNotEmpty &&
+              _effectiveController.baseOffset > 0 &&
+              _effectiveController.text
                   .substring(
                     0,
-                    messageInputController.baseOffset,
+                    _effectiveController.baseOffset,
                   )
                   .contains(':'),
           widget: _buildEmojiOverlay(),
@@ -553,7 +590,7 @@ class MessageInputState extends State<MessageInput> {
             height: 16,
             width: 16,
             foregroundDecoration: BoxDecoration(
-              border: messageInputController.showInChannel
+              border: _effectiveController.showInChannel
                   ? null
                   : Border.all(
                       color: _streamChatTheme.colorTheme.textHighEmphasis
@@ -565,20 +602,18 @@ class MessageInputState extends State<MessageInput> {
             child: Center(
               child: Material(
                 borderRadius: BorderRadius.circular(3),
-                color: messageInputController.showInChannel
+                color: _effectiveController.showInChannel
                     ? _streamChatTheme.colorTheme.accentPrimary
                     : _streamChatTheme.colorTheme.barsBg,
                 child: InkWell(
                   onTap: () {
-                    setState(() {
-                      messageInputController.showInChannel =
-                          !messageInputController.showInChannel;
-                    });
+                    _effectiveController.showInChannel =
+                        !_effectiveController.showInChannel;
                   },
                   child: AnimatedCrossFade(
                     duration: const Duration(milliseconds: 300),
                     reverseDuration: const Duration(milliseconds: 300),
-                    crossFadeState: messageInputController.showInChannel
+                    crossFadeState: _effectiveController.showInChannel
                         ? CrossFadeState.showFirst
                         : CrossFadeState.showSecond,
                     firstChild: StreamSvgIcon.check(
@@ -609,14 +644,14 @@ class MessageInputState extends State<MessageInput> {
 
   Widget _buildSendButton(BuildContext context) {
     if (widget.sendButtonBuilder != null) {
-      return widget.sendButtonBuilder!(context, messageInputController);
+      return widget.sendButtonBuilder!(context, _effectiveController);
     }
 
     return StreamMessageSendButton(
       onSendMessage: sendMessage,
       timeOut: _timeOut,
-      isIdle: !_messageIsPresent && messageInputController.attachments.isEmpty,
-      isEditEnabled: widget.editMessage != null,
+      isIdle: !widget.validator(_effectiveController.message),
+      isEditEnabled: _isEditing,
       idleSendButton: widget.idleSendButton,
       activeSendButton: widget.activeSendButton,
     );
@@ -663,7 +698,7 @@ class MessageInputState extends State<MessageInput> {
                   if (!widget.disableAttachments)
                     _buildAttachmentButton(context),
                   if (widget.showCommandsButton &&
-                      widget.editMessage == null &&
+                      !_isEditing &&
                       channel.state != null &&
                       channel.config?.commands.isNotEmpty == true)
                     _buildCommandButton(context),
@@ -714,7 +749,7 @@ class MessageInputState extends State<MessageInput> {
                     maxLines: null,
                     onSubmitted: (_) => sendMessage(),
                     keyboardType: widget.keyboardType,
-                    controller: messageInputController,
+                    controller: _effectiveController,
                     focusNode: _focusNode,
                     style: _messageInputTheme.inputTextStyle,
                     autofocus: widget.autofocus,
@@ -786,7 +821,7 @@ class MessageInputState extends State<MessageInput> {
                           size: 16,
                         ),
                         Text(
-                          _chosenCommand?.name.toUpperCase() ?? '',
+                          _effectiveController.value.command!.toUpperCase(),
                           style:
                               _streamChatTheme.textTheme.footnoteBold.copyWith(
                             color: Colors.white,
@@ -820,9 +855,7 @@ class MessageInputState extends State<MessageInput> {
                   height: 24,
                   width: 24,
                 ),
-                onPressed: () {
-                  setState(() => _commandEnabled = false);
-                },
+                onPressed: _effectiveController.clear,
               ),
             ),
           if (!_commandEnabled &&
@@ -837,14 +870,16 @@ class MessageInputState extends State<MessageInput> {
 
   late final _onChangedDebounced = debounce(
     () {
-      var value = messageInputController.text;
+      var value = _effectiveController.text;
       if (!mounted) return;
       value = value.trim();
 
       final channel = StreamChannel.of(context).channel;
       if (value.isNotEmpty) {
-        // ignore: no-empty-block
-        channel.keyStroke(widget.parentMessage?.id).catchError((e) {});
+        channel
+            .keyStroke(_effectiveController.value.parentId)
+            // ignore: no-empty-block
+            .catchError((e) {});
       }
 
       var actionsLength = widget.actions.length;
@@ -864,10 +899,10 @@ class MessageInputState extends State<MessageInput> {
   );
 
   String _getHint(BuildContext context) {
-    if (_commandEnabled && _chosenCommand!.name == 'giphy') {
+    if (_commandEnabled && _effectiveController.value.command == 'giphy') {
       return context.translations.searchGifLabel;
     }
-    if (messageInputController.attachments.isNotEmpty) {
+    if (_effectiveController.attachments.isNotEmpty) {
       return context.translations.addACommentOrSendLabel;
     }
     if (_timeOut != 0) {
@@ -879,16 +914,16 @@ class MessageInputState extends State<MessageInput> {
 
   void _checkEmoji(String s, BuildContext context) {
     if (s.isNotEmpty &&
-        messageInputController.baseOffset > 0 &&
-        messageInputController.text
+        _effectiveController.baseOffset > 0 &&
+        _effectiveController.text
             .substring(
               0,
-              messageInputController.baseOffset,
+              _effectiveController.baseOffset,
             )
             .contains(':')) {
-      final textToSelection = messageInputController.text.substring(
+      final textToSelection = _effectiveController.text.substring(
         0,
-        messageInputController.selectionStart,
+        _effectiveController.selectionStart,
       );
       final splits = textToSelection.split(':');
       final query = splits[splits.length - 2].toLowerCase();
@@ -902,11 +937,11 @@ class MessageInputState extends State<MessageInput> {
 
   void _checkMentions(String s, BuildContext context) {
     if (s.isNotEmpty &&
-        messageInputController.baseOffset > 0 &&
-        messageInputController.text
+        _effectiveController.baseOffset > 0 &&
+        _effectiveController.text
             .substring(
               0,
-              messageInputController.baseOffset,
+              _effectiveController.baseOffset,
             )
             .split(' ')
             .last
@@ -943,7 +978,7 @@ class MessageInputState extends State<MessageInput> {
   }
 
   Widget _buildCommandsOverlayEntry() {
-    final text = messageInputController.text.trimLeft();
+    final text = _effectiveController.text.trimLeft();
 
     final renderObject = context.findRenderObject() as RenderBox?;
     if (renderObject == null) {
@@ -959,7 +994,7 @@ class MessageInputState extends State<MessageInput> {
 
   Widget _buildFilePickerSection() {
     final picker = StreamAttachmentPicker(
-      messageInputController: messageInputController,
+      messageInputController: _effectiveController,
       onFilePicked: pickFile,
       isOpen: _openFilePickerSection,
       pickerSize: _openFilePickerSection ? _kMinMediaPickerSize : 0,
@@ -968,18 +1003,13 @@ class MessageInputState extends State<MessageInput> {
       maxAttachmentSize: widget.maxAttachmentSize,
       compressedVideoQuality: widget.compressedVideoQuality,
       compressedVideoFrameRate: widget.compressedVideoFrameRate,
-      onChangeInputState: (val) {
-        setState(() {
-          _inputEnabled = val;
-        });
-      },
       onError: _showErrorAlert,
     );
 
     if (_openFilePickerSection && widget.attachmentsPickerBuilder != null) {
       return widget.attachmentsPickerBuilder!(
         context,
-        messageInputController,
+        _effectiveController,
         picker,
       );
     }
@@ -988,14 +1018,15 @@ class MessageInputState extends State<MessageInput> {
   }
 
   Widget _buildMentionsOverlayEntry() {
-    if (messageInputController.selectionStart < 0) {
+    final channel = StreamChannel.of(context).channel;
+    if (_effectiveController.selectionStart < 0 || channel.state == null) {
       return const Offstage();
     }
 
-    final splits = messageInputController.text
+    final splits = _effectiveController.text
         .substring(
           0,
-          messageInputController.selectionStart,
+          _effectiveController.selectionStart,
         )
         .split('@');
     final query = splits.last.toLowerCase();
@@ -1020,23 +1051,19 @@ class MessageInputState extends State<MessageInput> {
       query: query,
       mentionAllAppUsers: widget.mentionAllAppUsers,
       client: StreamChat.of(context).client,
-      channel: StreamChannel.of(context).channel,
+      channel: channel,
       size: Size(renderObject.size.width - 16, 400),
       mentionsTileBuilder: tileBuilder,
       onMentionUserTap: (user) {
-        messageInputController.addMentionedUser(user);
+        _effectiveController.addMentionedUser(user);
         splits[splits.length - 1] = user.name;
         final rejoin = splits.join('@');
 
-        messageInputController.textEditingController.value = TextEditingValue(
-          text: rejoin +
-              messageInputController.text.substring(
-                messageInputController.selectionStart,
-              ),
-          selection: TextSelection.collapsed(
-            offset: rejoin.length,
-          ),
-        );
+        _effectiveController.text = rejoin +
+            _effectiveController.text.substring(
+              _effectiveController.selectionStart,
+            );
+
         _onChangedDebounced.cancel();
         setState(() => _showMentionsOverlay = false);
       },
@@ -1044,14 +1071,14 @@ class MessageInputState extends State<MessageInput> {
   }
 
   Widget _buildEmojiOverlay() {
-    if (messageInputController.baseOffset < 0) {
+    if (_effectiveController.baseOffset < 0) {
       return const Offstage();
     }
 
-    final splits = messageInputController.text
+    final splits = _effectiveController.text
         .substring(
           0,
-          messageInputController.baseOffset,
+          _effectiveController.baseOffset,
         )
         .split(':');
 
@@ -1071,22 +1098,17 @@ class MessageInputState extends State<MessageInput> {
   void _chooseEmoji(List<String> splits, Emoji emoji) {
     final rejoin = splits.sublist(0, splits.length - 1).join(':') + emoji.char!;
 
-    messageInputController.textEditingController.value = TextEditingValue(
-      text: rejoin +
-          messageInputController.text.substring(
-            messageInputController.selectionStart,
-          ),
-      selection: TextSelection.collapsed(
-        offset: rejoin.length,
-      ),
-    );
+    _effectiveController.text = rejoin +
+        _effectiveController.text.substring(
+          _effectiveController.selectionStart,
+        );
   }
 
   void _setCommand(Command c) {
-    messageInputController.clear();
+    _effectiveController
+      ..clear()
+      ..command = c;
     setState(() {
-      _chosenCommand = c;
-      _commandEnabled = true;
       _showCommandsOverlay = false;
     });
   }
@@ -1105,11 +1127,11 @@ class MessageInputState extends State<MessageInput> {
   }
 
   Widget _buildAttachments() {
-    if (messageInputController.attachments.isEmpty) return const Offstage();
-    final fileAttachments = messageInputController.attachments
+    if (_effectiveController.attachments.isEmpty) return const Offstage();
+    final fileAttachments = _effectiveController.attachments
         .where((it) => it.type == 'file')
         .toList(growable: false);
-    final remainingAttachments = messageInputController.attachments
+    final remainingAttachments = _effectiveController.attachments
         .where((it) => it.type != 'file')
         .toList(growable: false);
     return Column(
@@ -1127,9 +1149,7 @@ class MessageInputState extends State<MessageInput> {
                       (e) => ClipRRect(
                         borderRadius: BorderRadius.circular(10),
                         child: FileAttachment(
-                          message: Message(
-                            status: MessageSendingStatus.sending,
-                          ), // dummy message
+                          message: Message(), // dummy message
                           attachment: e,
                           size: Size(
                             MediaQuery.of(context).size.width * 0.65,
@@ -1196,9 +1216,10 @@ class MessageInputState extends State<MessageInput> {
           focusElevation: 0,
           hoverElevation: 0,
           onPressed: () {
-            setState(
-              () => messageInputController.attachments
-                  .removeWhere((e) => e.id == attachment.id),
+            _effectiveController.value = _effectiveController.value.copyWith(
+              attachments: _effectiveController.attachments
+                  .where((it) => it.id != attachment.id)
+                  .toList(),
             );
           },
           fillColor:
@@ -1278,7 +1299,7 @@ class MessageInputState extends State<MessageInput> {
   }
 
   Widget _buildCommandButton(BuildContext context) {
-    final s = messageInputController.text.trim();
+    final s = _effectiveController.text.trim();
     final defaultButton = IconButton(
       icon: StreamSvgIcon.lightning(
         color: s.isNotEmpty
@@ -1406,14 +1427,13 @@ class MessageInputState extends State<MessageInput> {
   ///
   /// Note: Only meant to be used from outside the state.
   void addAttachment(Attachment attachment) {
-    setState(() => _addAttachments([attachment]));
+    _addAttachments([attachment]);
   }
 
   /// Adds an attachment to the [messageInputController.attachments] map
   void _addAttachments(Iterable<Attachment> attachments) {
     final limit = widget.attachmentLimit;
-    final length =
-        messageInputController.attachments.length + attachments.length;
+    final length = _effectiveController.attachments.length + attachments.length;
     if (length > limit) {
       final onAttachmentLimitExceed = widget.onAttachmentLimitExceed;
       if (onAttachmentLimitExceed != null) {
@@ -1427,7 +1447,7 @@ class MessageInputState extends State<MessageInput> {
       );
     }
     for (final attachment in attachments) {
-      messageInputController.addAttachment(attachment);
+      _effectiveController.addAttachment(attachment);
     }
   }
 
@@ -1476,7 +1496,6 @@ class MessageInputState extends State<MessageInput> {
       }
       final res = await FilePicker.platform.pickFiles(
         type: type,
-        withData: true,
       );
       if (res?.files.isNotEmpty == true) {
         file = res!.files.single.toAttachmentFile;
@@ -1534,69 +1553,25 @@ class MessageInputState extends State<MessageInput> {
       }
     }
 
-    setState(() {
-      _addAttachments([
-        attachment.copyWith(
-          file: file,
-          extraData: {...attachment.extraData}
-            ..update('file_size', ((_) => file!.size!)),
-        ),
-      ]);
-    });
+    _addAttachments([
+      attachment.copyWith(
+        file: file,
+        extraData: {...attachment.extraData}
+          ..update('file_size', ((_) => file!.size!)),
+      ),
+    ]);
   }
 
   /// Sends the current message
   Future<void> sendMessage() async {
-    var text = messageInputController.text.trim();
-    final attachments = messageInputController.attachments;
+    var message = _effectiveController.value;
 
-    if (text.isEmpty && attachments.isEmpty) {
-      return;
-    }
+    var shouldKeepFocus = widget.shouldKeepFocusAfterMessage;
 
-    final shouldUnfocus = _commandEnabled;
+    shouldKeepFocus ??= !_commandEnabled;
 
-    if (_commandEnabled) {
-      text = '${'/${_chosenCommand!.name} '}$text';
-    }
-
-    messageInputController
-      ..text = ''
-      ..clearAttachments();
+    _effectiveController.reset();
     widget.onQuotedMessageCleared?.call();
-
-    setState(() {
-      _commandEnabled = false;
-    });
-
-    Message message;
-    if (widget.editMessage != null) {
-      message = widget.editMessage!.copyWith(
-        text: text,
-        attachments: attachments,
-        mentionedUsers: messageInputController.mentionedUsers
-            .where((u) => text.contains('@${u.name}'))
-            .toList(),
-      );
-    } else {
-      message = (widget.initialMessage ?? Message()).copyWith(
-        parentId: widget.parentMessage?.id,
-        text: text,
-        attachments: attachments,
-        mentionedUsers: messageInputController.mentionedUsers
-            .where((u) => text.contains('@${u.name}'))
-            .toList(),
-        showInChannel: widget.parentMessage != null
-            ? messageInputController.showInChannel
-            : null,
-      );
-    }
-
-    if (widget.quotedMessage != null) {
-      message = message.copyWith(
-        quotedMessageId: widget.quotedMessage!.id,
-      );
-    }
 
     if (widget.preMessageSending != null) {
       message = await widget.preMessageSending!(message);
@@ -1608,25 +1583,23 @@ class MessageInputState extends State<MessageInput> {
       await streamChannel.reloadChannel();
     }
 
-    messageInputController.clearMentionedUsers();
-
     try {
       Future sendingFuture;
-      if (widget.editMessage == null ||
-          widget.editMessage!.status == MessageSendingStatus.failed ||
-          widget.editMessage!.status == MessageSendingStatus.sending) {
+      if (!_isEditing) {
         sendingFuture = channel.sendMessage(message);
       } else {
         sendingFuture = channel.updateMessage(message);
       }
 
-      if (!shouldUnfocus) {
+      if (shouldKeepFocus) {
         FocusScope.of(context).requestFocus(_focusNode);
+      } else {
+        FocusScope.of(context).unfocus();
       }
 
       final resp = await sendingFuture;
       if (resp.message?.type == 'error') {
-        _parseExistingMessage(message);
+        _effectiveController.value = message;
       }
       _startSlowMode();
       widget.onMessageSent?.call(resp.message);
@@ -1705,36 +1678,23 @@ class MessageInputState extends State<MessageInput> {
     );
   }
 
-  void _parseExistingMessage(Message message) {
-    final messageText = message.text;
-    if (messageText != null) messageInputController.text = messageText;
-    _addAttachments(message.attachments);
-  }
-
   @override
   void dispose() {
-    messageInputController.textEditingController
+    _effectiveController.textEditingController
         .removeListener(_onChangedDebounced);
-    messageInputController.dispose();
+    _controller?.dispose();
     _focusNode.removeListener(_focusNodeListener);
     _stopSlowMode();
     _onChangedDebounced.cancel();
     super.dispose();
   }
 
-  bool _initialized = false;
-
   @override
   void didChangeDependencies() {
     _streamChatTheme = StreamChatTheme.of(context);
     _messageInputTheme = MessageInputTheme.of(context);
-    if (widget.editMessage == null && _timeOut <= 0) _startSlowMode();
+    if (!_isEditing && _timeOut <= 0) _startSlowMode();
 
-    if ((widget.editMessage != null || widget.initialMessage != null) &&
-        !_initialized) {
-      FocusScope.of(context).requestFocus(_focusNode);
-      _initialized = true;
-    }
     super.didChangeDependencies();
   }
 }
