@@ -21,30 +21,41 @@ typedef CustomAttachmentIconBuilder = Widget Function(
   bool active,
 );
 
-///
+/// A widget that allows to pick an attachment.
 class StreamAttachmentPicker extends StatefulWidget {
+  /// True if the picker is open.
   final bool isOpen;
+
+  /// The picker size in height.
   final double pickerSize;
+
+  /// The [MessageInputController] linked to this picker.
   final MessageInputController messageInputController;
+
+  /// The limit of attachments that can be picked.
   final int attachmentLimit;
+
+  /// The callback for when the attachment limit is exceeded.
   final AttachmentLimitExceedListener? onAttachmentLimitExceeded;
-  final ValueChanged<bool>? onChangeInputState;
+
   final ValueChanged<String>? onError;
   final FilePickerCallback onFilePicked;
 
-  /// Video quality to use when compressing the videos
+  /// Video quality to use when compressing the videos.
   final VideoQuality compressedVideoQuality;
 
-  /// Frame rate to use when compressing the videos
+  /// Frame rate to use when compressing the videos.
   final int compressedVideoFrameRate;
 
-  /// Max attachment size in bytes
-  /// Defaults to 20 MB
-  /// do not set it if you're using our default CDN
+  /// Max attachment size in bytes:
+  /// - Defaults to 20 MB
+  /// - Do not set it if you're using our default CDN
   final int maxAttachmentSize;
 
+  /// The list of attachment types that can be picked.
   final List<DefaultAttachmentTypes> allowedAttachmentTypes;
 
+  /// The list of custom attachment types that can be picked.
   final List<CustomAttachmentType> customAttachmentTypes;
 
   const StreamAttachmentPicker({
@@ -58,7 +69,6 @@ class StreamAttachmentPicker extends StatefulWidget {
     this.maxAttachmentSize = 20971520,
     this.compressedVideoQuality = VideoQuality.DefaultQuality,
     this.compressedVideoFrameRate = 30,
-    this.onChangeInputState,
     this.onError,
     this.allowedAttachmentTypes = const [
       DefaultAttachmentTypes.image,
@@ -99,7 +109,6 @@ class StreamAttachmentPicker extends StatefulWidget {
             compressedVideoQuality ?? this.compressedVideoQuality,
         compressedVideoFrameRate:
             compressedVideoFrameRate ?? this.compressedVideoFrameRate,
-        onChangeInputState: onChangeInputState ?? this.onChangeInputState,
         onError: onError ?? this.onError,
         allowedAttachmentTypes:
             allowedAttachmentTypes ?? this.allowedAttachmentTypes,
@@ -116,8 +125,8 @@ class _StreamAttachmentPickerState extends State<StreamAttachmentPicker> {
 
   @override
   Widget build(BuildContext context) {
-    var _streamChatTheme = StreamChatTheme.of(context);
-    var messageInputController = widget.messageInputController;
+    final _streamChatTheme = StreamChatTheme.of(context);
+    final messageInputController = widget.messageInputController;
 
     final _attachmentContainsImage =
         messageInputController.attachments.any((it) => it.type == 'image');
