@@ -1,15 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
+/// A widget that displays a sending button.
 class StreamMessageSendButton extends StatelessWidget {
-  final int timeOut;
-  final bool isIdle;
-  final bool isCommandEnabled;
-  final bool isEditEnabled;
-  final Widget? idleSendButton;
-  final Widget? activeSendButton;
-  final VoidCallback onSendMessage;
-
+  /// Returns a [StreamMessageSendButton] with the given [timeOut], [isIdle],
+  /// [isCommandEnabled], [isEditEnabled], [idleSendButton], [activeSendButton],
+  /// [onSendMessage].
   const StreamMessageSendButton({
     Key? key,
     this.timeOut = 0,
@@ -21,9 +17,30 @@ class StreamMessageSendButton extends StatelessWidget {
     required this.onSendMessage,
   }) : super(key: key);
 
+  /// Time out related to slow mode.
+  final int timeOut;
+
+  /// If true the button will be disabled.
+  final bool isIdle;
+
+  /// True if a command is being sent.
+  final bool isCommandEnabled;
+
+  /// True if in editing mode.
+  final bool isEditEnabled;
+
+  /// The widget to display when the button is disabled.
+  final Widget? idleSendButton;
+
+  /// The widget to display when the button is enabled.
+  final Widget? activeSendButton;
+
+  /// The callback to call when the button is pressed.
+  final VoidCallback onSendMessage;
+
   @override
   Widget build(BuildContext context) {
-    var _streamChatTheme = StreamChatTheme.of(context);
+    final _streamChatTheme = StreamChatTheme.of(context);
 
     late Widget sendButton;
     if (timeOut > 0) {
@@ -46,7 +63,7 @@ class StreamMessageSendButton extends StatelessWidget {
   }
 
   Widget _buildIdleSendButton(BuildContext context) {
-    var _messageInputTheme = MessageInputTheme.of(context);
+    final _messageInputTheme = MessageInputTheme.of(context);
 
     return Padding(
       padding: const EdgeInsets.all(8),
@@ -58,7 +75,7 @@ class StreamMessageSendButton extends StatelessWidget {
   }
 
   Widget _buildSendButton(BuildContext context) {
-    var _messageInputTheme = MessageInputTheme.of(context);
+    final _messageInputTheme = MessageInputTheme.of(context);
 
     return Padding(
       padding: const EdgeInsets.all(8),
