@@ -2,6 +2,7 @@ export 'package:jiffy/jiffy.dart';
 export 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
 
 export 'src/attachment/attachment.dart';
+export 'src/attachment_actions_modal.dart';
 export 'src/back_button.dart';
 export 'src/channel_avatar.dart';
 export 'src/channel_header.dart';
@@ -22,16 +23,16 @@ export 'src/localization/stream_chat_localizations.dart';
 export 'src/localization/translations.dart' show DefaultTranslations;
 export 'src/mention_tile.dart';
 export 'src/message_action.dart';
-export 'src/message_input.dart';
+export 'src/message_input/message_input.dart';
 export 'src/message_list_view.dart';
 export 'src/message_search_item.dart';
 export 'src/message_search_list_view.dart';
 export 'src/message_text.dart';
 export 'src/message_widget.dart';
-export 'src/mip/countdown_button.dart';
-export 'src/mip/stream_attachment_picker.dart';
-export 'src/mip/stream_message_send_button.dart';
-export 'src/mip/stream_message_text_field.dart';
+export 'src/message_input/countdown_button.dart';
+export 'src/message_input/stream_attachment_picker.dart';
+export 'src/message_input/stream_message_send_button.dart';
+export 'src/message_input/stream_message_text_field.dart';
 export 'src/option_list_tile.dart';
 export 'src/reaction_icon.dart';
 export 'src/reaction_picker.dart';
