@@ -242,7 +242,7 @@ class _MessageViewState extends State<MessageView> {
                       ),
                     ),
                   ),
-                )
+                ),
               ],
             ),
           )
